requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.24.3
pydicom==2.4.3
Pillow==10.0.0